@@ -1,15 +1,17 @@
 import logging
 import os
 from dataclasses import dataclass
+from pathlib import Path
 from typing import Any, List, Optional
 
+import matplotlib.pyplot as plt
 import pandas as pd
-from sklearn.model_selection import StratifiedKFold, train_test_split
+from sklearn.model_selection import train_test_split
 
 from autorad.config import config
 from autorad.config.type_definitions import PathLike
-from autorad.utils import io, utils
-from autorad.utils.splitting import split_full_dataset
+from autorad.utils import io, splitting, utils
+from autorad.visualization import matplotlib_utils, plot_volumes
 
 log = logging.getLogger(__name__)
 
@@ -50,6 +52,22 @@ class TrainingData:
     _X_preprocessed: Optional[TrainingInput] = None
     _y_preprocessed: Optional[TrainingLabels] = None
 
+    def iter_training(self):
+        X, y = self._X_preprocessed, self._y_preprocessed
+        if X is None:
+            raise ValueError("No preprocessing done!")
+        if X.val is not None:
+            yield X.train, y.train, X.val, y.val
+        elif self.X_preprocessed.train_folds is not None:
+            yield from zip(
+                X.train_folds, y.train_folds, X.val_folds, y.val_folds
+            )
+
+    def __repr__(self):
+        return f"TrainingData with {len(self.y.train)} training observations,\
+        {len(self.y.test)} test observations, {self.X.train.shape[1]} features \
+        and {self.meta.train.shape[1]} meta columns."
+
     @property
     def X_preprocessed(self):
         if self._X_preprocessed is None:
@@ -65,8 +83,7 @@ class TrainingData:
 
 class FeatureDataset:
     """
-    Store the extarcted features and labels, split into training/test sets,
-    select features and show them.
+    Store the extracted features and labels, split into training/test sets.
     """
 
     def __init__(
@@ -74,8 +91,8 @@ class FeatureDataset:
         dataframe: pd.DataFrame,
         target: str,
         ID_colname: str,
-        features: Optional[List[str]] = None,
-        meta_columns: List[str] = [],
+        features: Optional[list[str]] = None,
+        meta_columns: list[str] = [],
         random_state: int = config.SEED,
     ):
         """
@@ -85,9 +102,7 @@ class FeatureDataset:
             ID_colname: name of column with unique IDs for each case
             features: feature names
             meta_columns: columns to keep that are not features
-            random_state: random seed
-        Returns:
-            None
+            random_state: seed, used for splitting
         """
         self.df = dataframe
         self.target = target
@@ -99,8 +114,6 @@ class FeatureDataset:
         self.meta_df = self.df[meta_columns + [ID_colname]]
         self._data: Optional[TrainingData] = None
         self.cv_splits: Optional[List[tuple[Any, Any]]] = None
-        self.selected_features: Optional[List[str]] = None
-        self.result_dir = config.RESULT_DIR
 
     def _init_features(
         self, features: Optional[List[str]] = None
@@ -134,55 +147,89 @@ class FeatureDataset:
             split_on = self.ID_colname
         test_ids = splits["test"]
         test_rows = self.df[split_on].isin(test_ids)
-        train_rows = ~self.df[split_on].isin(test_ids)
 
         # Split dataframe rows
         X, y, meta = {}, {}, {}
+
         X["test"] = self.X.loc[test_rows]
         y["test"] = self.y.loc[test_rows]
         meta["test"] = self.meta_df.loc[test_rows]
+
+        train_rows = ~self.df[split_on].isin(test_ids)
+
+        if "val" in splits:
+            val_ids = splits["val"]
+            val_rows = self.df[split_on].isin(val_ids)
+            train_rows = train_rows & ~val_rows
+
+            X["val"] = self.X.loc[val_rows]
+            y["val"] = self.y.loc[val_rows]
+            meta["val"] = self.meta_df.loc[val_rows]
+
         X["train"] = self.X.loc[train_rows]
         y["train"] = self.y.loc[train_rows]
         meta["train"] = self.meta_df.loc[train_rows]
 
-        train_ids = splits["train"]
-        n_splits = len(train_ids)
-        self.cv_splits = [
-            (train_ids[f"fold_{i}"][0], train_ids[f"fold_{i}"][1])
-            for i in range(n_splits)
-        ]
-        X["train_folds"], X["val_folds"] = [], []
-        y["train_folds"], y["val_folds"] = [], []
-        meta["train_folds"], meta["val_folds"] = [], []
-        for train_fold_ids, val_fold_ids in self.cv_splits:
+        if isinstance(splits["train"], dict):
+            train_splits = splits["train"]
+            n_splits = len(train_splits)
+            self.cv_splits = [
+                train_splits[f"fold_{i}"] for i in range(n_splits)
+            ]
+            X["train_folds"], X["val_folds"] = [], []
+            y["train_folds"], y["val_folds"] = [], []
+            meta["train_folds"], meta["val_folds"] = [], []
+            for train_fold_ids, val_fold_ids in self.cv_splits:
 
-            train_fold_rows = self.df[split_on].isin(train_fold_ids)
-            val_fold_rows = self.df[split_on].isin(val_fold_ids)
+                train_fold_rows = self.df[split_on].isin(train_fold_ids)
+                val_fold_rows = self.df[split_on].isin(val_fold_ids)
 
-            X["train_folds"].append(self.X[train_fold_rows])
-            X["val_folds"].append(self.X[val_fold_rows])
-            y["train_folds"].append(self.y[train_fold_rows])
-            y["val_folds"].append(self.y[val_fold_rows])
-            meta["train_folds"].append(self.meta_df[train_fold_rows])
-            meta["val_folds"].append(self.meta_df[val_fold_rows])
+                X["train_folds"].append(self.X[train_fold_rows])
+                X["val_folds"].append(self.X[val_fold_rows])
+                y["train_folds"].append(self.y[train_fold_rows])
+                y["val_folds"].append(self.y[val_fold_rows])
+                meta["train_folds"].append(self.meta_df[train_fold_rows])
+                meta["val_folds"].append(self.meta_df[val_fold_rows])
         self._data = TrainingData(
             TrainingInput(**X), TrainingLabels(**y), TrainingMeta(**meta)
         )
         return self
 
-    def full_split(
+    def split(
         self,
         save_path: PathLike,
+        method="train_with_cross_validation_test",
         split_on: Optional[str] = None,
+        test_size: float = 0.2,
+        *args,
+        **kwargs,
+    ):
+        if split_on is None:
+            split_on = self.ID_colname
+        if method == "train_with_cross_validation_test":
+            splits = self.full_split(split_on, test_size, *args, **kwargs)
+        elif method == "train_val_test":
+            splits = self.split_train_val_test(
+                split_on, test_size, *args, **kwargs
+            )
+        else:
+            raise ValueError(f"Method {method} is not supported.")
+
+        Path(save_path).parent.mkdir(parents=True, exist_ok=True)
+        io.save_json(splits, save_path)
+        self.load_splits_from_json(save_path)
+        return splits
+
+    def full_split(
+        self,
+        split_on: str,
         test_size: float = 0.2,
         n_splits: int = 5,
     ):
         """
-        Split into test and training, split training into 5 folds.
+        Split into test and training, split training into k folds.
         Save the splits to json.
         """
-        if split_on is None:
-            split_on = self.ID_colname
         patient_df = self.df[[split_on, self.target]].drop_duplicates()
         if not patient_df[split_on].is_unique:
             raise ValueError(
@@ -190,13 +237,29 @@ class FeatureDataset:
             )
         ids = patient_df[split_on].tolist()
         labels = patient_df[self.target].tolist()
-        split_full_dataset(
+        splits = splitting.split_full_dataset(
             ids=ids,
             labels=labels,
-            save_path=save_path,
             test_size=test_size,
             n_splits=n_splits,
         )
+        return splits
+
+    def split_train_val_test(
+        self,
+        split_on: str,
+        test_size: float = 0.2,
+        val_size: float = 0.2,
+    ):
+        ids = self.df[split_on].tolist()
+        labels = self.df[self.target].tolist()
+        splits = splitting.split_train_val_test(
+            ids=ids,
+            labels=labels,
+            test_size=test_size,
+            val_size=val_size,
+        )
+        return splits
 
     def get_train_test_split_from_column(
         self, column_name: str, test_value: str
@@ -211,7 +274,7 @@ class FeatureDataset:
 
         return X_train, y_train, X_test, y_test
 
-    def split_dataset_test_from_column(
+    def split_train_val_test_from_column(
         self, column_name: str, test_value: str, val_size: float = 0.2
     ):
         data = {}
@@ -234,50 +297,51 @@ class FeatureDataset:
         )
         self._data = TrainingData(**data)
 
-    def split_cross_validation(self, X, y, n_splits: int):
-        """
-        Split using stratified k-fold cross-validation.
-        """
-        kf = StratifiedKFold(
-            n_splits=n_splits, shuffle=True, random_state=self.random_state
-        )
-        cv_split_generator = kf.split(X, y)
-        self.cv_splits = list(cv_split_generator)
-        self._add_cross_validation_folds(X, y)
-        return self
-
-    def _add_cross_validation_folds(self, X, y):
-        X_train_fold, y_train_fold = [], []
-        X_val_fold, y_val_fold = [], []
-        if self.cv_splits is None:
-            raise ValueError("no cross-validation split available!")
-        for _, (train_idx, val_idx) in enumerate(self.cv_splits):
-            X_train_fold.append(X.iloc[train_idx])
-            y_train_fold.append(y.iloc[train_idx])
-
-            X_val_fold.append(X.iloc[val_idx])
-            y_val_fold.append(y.iloc[val_idx])
-
-        return self
-
-    def split_cross_validation_test_from_column(
-        self, column_name: str, test_value: str, n_splits: int = 5
+    def full_split_with_test_from_column(
+        self,
+        column_name: str,
+        test_value: str,
+        save_path: PathLike,
+        split_on: Optional[str] = None,
+        split_label: Optional[str] = None,
+        n_splits: int = 5,
     ):
         """
         Splits into train and test according to `column_name`,
         then performs stratified k-fold cross validation split
         on the training set.
         """
-        data = {}
-        (
-            data["X_train"],
-            data["y_train"],
-            data["X_test"],
-            data["y_test"],
-        ) = self.get_train_test_split_from_column(column_name, test_value)
-        data = self.split_cross_validation(
-            data["X_train"], data["y_train"], n_splits=n_splits
+        if split_on is None:
+            split_on = self.ID_colname
+        if split_label is None:
+            split_label = self.target
+        df_to_split = self.df[
+            [split_on, split_label, column_name]
+        ].drop_duplicates()
+        if not df_to_split[split_on].is_unique:
+            raise ValueError(
+                f"Selected column {split_on} has varying labels for the same ID!"
+            )
+        train_to_split = df_to_split[df_to_split[column_name] != test_value]
+        ids_train = train_to_split[split_on].tolist()
+        y_train = train_to_split[split_label].tolist()
+        ids_test = df_to_split.loc[
+            df_to_split[column_name] == test_value, split_on
+        ].tolist()
+
+        ids_train_cv = splitting.split_cross_validation(
+            ids_train, y_train, n_splits, random_state=self.random_state
         )
+
+        ids_split = {
+            "split_type": f"predefined test as {column_name} = {test_value}"
+            " and stratified cross validation on training",
+            "test": ids_test,
+            "train": ids_train_cv,
+        }
+        io.save_json(ids_split, save_path)
+        self.load_splits_from_json(save_path)
+
         return self
 
 
@@ -289,8 +353,8 @@ class ImageDataset:
     def __init__(
         self,
         df: pd.DataFrame,
-        image_colname: str,
-        mask_colname: str,
+        image_colname: str = "image_path",
+        mask_colname: str = "segmentation_path",
         ID_colname: Optional[str] = None,
         root_dir: Optional[PathLike] = None,
     ):
@@ -299,37 +363,38 @@ class ImageDataset:
             df: dataframe with image and mask paths
             image_colname: name of the image column in df
             mask_colname: name of the mask column in df
-            ID_colname: name of the ID column in df, if not given,
+            ID_colname: name of the ID column in df. If None,
                 IDs are assigned sequentially
             root_dir: root directory of the dataset, if needed
                 to resolve paths
-        Returns:
-            None
         """
-        self.df = df
+        self._df = df
         self.image_colname = self._check_if_in_df(image_colname)
         self.mask_colname = self._check_if_in_df(mask_colname)
         self._set_ID_col(ID_colname)
         self.root_dir = root_dir
 
     def _check_if_in_df(self, colname: str):
-        if colname not in self.df.columns:
+        if colname not in self._df.columns:
             raise ValueError(
                 f"{colname} not found in columns of the dataframe."
             )
-        if self.df[colname].isnull().any():
+        if self._df[colname].isnull().any():
             raise ValueError(f"{colname} contains null values")
         return colname
 
     def _set_new_IDs(self):
         log.info("ID not set. Assigning sequential IDs.")
-        self.ID_colname = "ID"
-        self.df[self.ID_colname] = self.df.index
+        if "ID" not in self._df.columns:
+            self.ID_colname = "ID"
+        else:
+            self.ID_colname = "ID_autogenerated"
+        self._df[self.ID_colname] = range(len(self._df))
 
     def _set_ID_col_from_given(self, id_colname: str):
-        if id_colname not in self.df.columns:
+        if id_colname not in self._df.columns:
             raise ValueError(f"{id_colname} not in columns of dataframe.")
-        ids = self.df[id_colname]
+        ids = self._df[id_colname]
         # assert IDs are unique
         if len(ids.unique()) != len(ids):
             raise ValueError("IDs are not unique!")
@@ -341,30 +406,60 @@ class ImageDataset:
         else:
             self._set_ID_col_from_given(id_colname)
 
-    def get_df(self) -> pd.DataFrame:
+    @property
+    def df(self) -> pd.DataFrame:
+        """If root_dir is set, returns the dataframe with paths resolved"""
         if self.root_dir is None:
-            return self.df
-        return self.df.assign(
+            return self._df
+        result = self._df.copy()
+        return result.assign(
             **{
-                self.image_colname: self.df[self.image_colname].apply(
+                self.image_colname: self._df[self.image_colname].apply(
                     lambda x: os.path.join(self.root_dir, x)
                 )
             }
         ).assign(
             **{
-                self.mask_colname: self.df[self.mask_colname].apply(
+                self.mask_colname: self._df[self.mask_colname].apply(
                     lambda x: os.path.join(self.root_dir, x)
                 )
             }
         )
 
+    @property
     def image_paths(self) -> List[str]:
         return self.df[self.image_colname].to_list()
 
+    @property
     def mask_paths(self) -> List[str]:
         return self.df[self.mask_colname].to_list()
 
+    @property
     def ids(self) -> List[str]:
         if self.ID_colname is None:
             raise AttributeError("ID is not set.")
         return self.df[self.ID_colname].to_list()
+
+    def plot_examples(self, n: int = 1, window="soft tissues"):
+        if n > len(self.image_paths):
+            n = len(self.image_paths)
+            log.info(
+                f"Not enough cases. Plotting all the cases instead (n={n})"
+            )
+        df_to_plot = self.df.sample(n)
+        nrows, ncols, figsize = matplotlib_utils.get_subplots_dimensions(n)
+        fig, axs = plt.subplots(nrows, ncols, figsize=figsize)
+        for i in range(len(df_to_plot)):
+            case = df_to_plot.iloc[i]
+            ax = axs.flat[i]
+            vols = plot_volumes.BaseVolumes.from_nifti(
+                case[self.image_colname],
+                case[self.mask_colname],
+                window=window,
+            )
+            image_2D, mask_2D = vols.get_slices()
+            single_plot = plot_volumes.overlay_mask_contour(image_2D, mask_2D)
+            ax.imshow(single_plot)
+            ax.set_title(f"{case[self.ID_colname]}")
+            ax.axis("off")
+        # return fig
