@@ -1,7 +1,12 @@
 import json
 import logging
+import logging.config
 import os
+import sys
 import tempfile
+from pathlib import Path
+
+from rich.logging import RichHandler
 
 import autorad
 
@@ -38,7 +43,7 @@ AVAILABLE_CLASSIFIERS = [
     "SVM",
     "XGBoost",
 ]
-FEATURE_SELECTION_METHODS = ["anova", "lasso", "boruta"]
+FEATURE_SELECTION_METHODS = ["anova", "lasso", "boruta", "boruta-shap"]
 OVERSAMPLING_METHODS = ["SMOTE", "ADASYN", None]
 
 SEED = 123
@@ -49,4 +54,49 @@ PYRADIOMICS_TMP_DIR = tempfile.mkdtemp()
 IS_DEMO = False
 
 # Logging
-logging.getLogger("autorad").addHandler(logging.NullHandler())
+LOGS_DIR = Path(RESULT_DIR, "logs")
+LOGS_DIR.mkdir(parents=True, exist_ok=True)
+
+logging_config = {
+    "version": 1,
+    "disable_existing_loggers": False,
+    "formatters": {
+        "minimal": {"format": "%(message)s"},
+        "detailed": {
+            "format": "%(levelname)s %(asctime)s [%(name)s:%(filename)s:%(funcName)s:%(lineno)d]\n%(message)s\n"
+        },
+    },
+    "handlers": {
+        "console": {
+            "class": "logging.StreamHandler",
+            "stream": sys.stdout,
+            "formatter": "minimal",
+            "level": logging.DEBUG,
+        },
+        "info": {
+            "class": "logging.handlers.RotatingFileHandler",
+            "filename": Path(LOGS_DIR, "info.log"),
+            "maxBytes": 10485760,  # 1 MB
+            "backupCount": 10,
+            "formatter": "detailed",
+            "level": logging.INFO,
+        },
+        "error": {
+            "class": "logging.handlers.RotatingFileHandler",
+            "filename": Path(LOGS_DIR, "error.log"),
+            "maxBytes": 10485760,  # 1 MB
+            "backupCount": 10,
+            "formatter": "detailed",
+            "level": logging.ERROR,
+        },
+    },
+    "root": {
+        "handlers": ["console", "info", "error"],
+        "level": logging.INFO,
+        "propagate": True,
+    },
+}
+
+logging.config.dictConfig(logging_config)
+logger = logging.getLogger()
+logger.handlers[0] = RichHandler(markup=True)
