@@ -4,8 +4,6 @@ import numpy as np
 from sklearn.model_selection import StratifiedKFold, train_test_split
 
 from autorad.config import config
-from autorad.config.type_definitions import PathLike
-from autorad.utils import io
 
 
 def split_cross_validation(
@@ -13,7 +11,7 @@ def split_cross_validation(
     labels: Sequence[Union[str, int]],
     n_splits: int = 5,
     random_state: int = config.SEED,
-):
+) -> dict:
     """
     Split data into n_splits folds for cross-validation, with stratification.
     """
@@ -32,11 +30,10 @@ def split_cross_validation(
 def split_full_dataset(
     ids: Sequence[str],
     labels: Sequence[str],
-    save_path: PathLike,
     test_size: float = 0.2,
     n_splits: int = 5,
     random_state: int = config.SEED,
-):
+) -> dict:
     """
     Split data into test and training, then divide training into n folds for
     cross-validation. Labels are needed for stratification.
@@ -53,22 +50,21 @@ def split_full_dataset(
         ids_train, y_train, n_splits, random_state=random_state
     )
     ids_split = {
-        "split_type": "test + cross-validation on the training",
+        "split_type": f"{test_size:.0%} test + {n_splits}-fold "
+        "cross-validation on the training set",
         "test": ids_test,
         "train": ids_train_cv,
     }
-    io.save_json(ids_split, save_path)
     return ids_split
 
 
 def split_train_val_test(
     ids: Sequence[str],
     labels: Sequence[str],
-    save_path: PathLike,
     test_size: float = 0.2,
     val_size: float = 0.2,
     random_state: int = config.SEED,
-):
+) -> dict:
     """
     Stratified train/val/test split without cross-validation.
     """
@@ -88,9 +84,12 @@ def split_train_val_test(
         stratify=y_train_val,
         random_state=random_state,
     )
-    ids_split = {}
-    ids_split["train"] = ids_train
-    ids_split["val"] = ids_val
-    ids_split["test"] = ids_test
-    io.save_json(ids_split, save_path)
+    train_size = 1 - val_size - test_size
+    ids_split = {
+        "split_type": f"stratified split: {train_size:.0%} train + {val_size:.0%} validation"
+        f" + {test_size:.0%} test",
+        "train": ids_train,
+        "val": ids_val,
+        "test": ids_test,
+    }
     return ids_split
