@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 
 import streamlit as st
@@ -13,8 +14,10 @@ pretrained_models = io.load_json(json_path)
 
 def show():
     with st.sidebar:
-        modalities = ["CT", "MRI"]
-        st.selectbox("Modality", modalities)
+        # modalities = ["CT", "MRI"]
+        # modality = st.selectbox("Modality", modalities)
+        # regions = ["abdomen", "chest", "pelvis", "head"]
+        # region = st.selectbox("Region", regions)
         organs = [
             "Brain",
             "Lung",
@@ -25,12 +28,27 @@ def show():
             "Pancreas",
             "Intestine",
         ]
-        st.selectbox("Organ", organs)
+        st.multiselect("Organs", organs)
 
+    input_dir = st.text_input("Path to the directory with images")
+    input_dir = Path(input_dir.strip('"'))
+    if not input_dir.is_dir():
+        st.error("Directory not found!")
+    files = list(input_dir.glob("*.nii.gz"))
+    if files:
+        st.success(f"{len(files)} images found!")
+    model_dim = st.selectbox("Model", ["2D", "3D"])
+    model = "2d" if model_dim == "2D" else "3d_fullres"
     # load the template
+    task = "Task017_AbdominalOrganSegmentation"
     env = Environment(loader=FileSystemLoader("webapp/templates/segmentation"))
     template = env.get_template("nnunet_code.py.jinja")
-    code = template.render(header=template_utils.code_header, notebook=False)
+    code = template.render(
+        header=template_utils.code_header,
+        notebook=False,
+        task=task,
+        model=model,
+    )
     notebook_code = template.render(
         header=template_utils.notebook_header, notebook=True
     )
@@ -46,7 +64,11 @@ def show():
         template_utils.download_button(
             notebook, "generated-notebook.ipynb", "📓 Download (.ipynb)"
         )
-
+    run_jupyter = st.button("📂 Open in Jupyter Notebook")
+    if run_jupyter:
+        with open("generated-notebook.ipynb", "w") as f:
+            f.write(notebook)
+        os.system("jupyter notebook")
     st.code(code)
 
 
